[REFACTOR] Move Norm and Activation to Modules.py
